refactor(oxlint-plugin): use shared enclosingComponentOrHookScope util in prefer-module-scope-* rules
Sister to the previous commit. The two architecture rules (
prefer-module-scope-pure-function and prefer-module-scope-static-value
) carried their own copies of the same enclosing-function finder, this
time scope-aware (returning the body ScopeDescriptor for closure-
capture queries).

Extracts the variant to utils/enclosing-component-or-hook-scope.ts
with an EnclosingComponentInfo interface that exposes functionNode,
bodyScope, and displayName. Both architecture rules now consume it.

Also drops an unused COMPONENT_SCOPE_KINDS constant that was left
behind during earlier development.

Addresses Bugbot's 'Four files duplicate component/hook enclosing-
function finder logic' finding.
